# server/openapi.yaml
openapi: 3.0.3
info:
  title: Agrocylo Backend API
  version: 1.0.0
paths:
  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: farmer
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200': { description: Product list }
    post:
      summary: Create product
      parameters:
        - in: header
          name: x-wallet-address
          required: true
          schema: { type: string }
      responses:
        '201': { description: Created }
  /products/{id}:
    get:
      summary: Product detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Product detail }
        '404': { description: Not found }
    patch:
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: x-wallet-address
          required: true
          schema: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      summary: Soft delete product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: x-wallet-address
          required: true
          schema: { type: string }
      responses:
        '200': { description: Soft deleted }
  /products/{product_id}/image:
    post:
      summary: Upload product image
      responses:
        '200': { description: Image uploaded }
        '413': { description: Payload too large }
        '415': { description: Unsupported media type }
    delete:
      summary: Delete product image
      responses:
        '204': { description: Deleted }
  /cart:
    get:
      summary: Get active cart
      parameters:
        - in: header
          name: x-wallet-address
          required: true
          schema: { type: string }
      responses:
        '200': { description: Active cart }
    delete:
      summary: Clear cart
      responses:
        '200': { description: Cart cleared }
  /cart/items:
    post:
      summary: Add item to cart
      responses:
        '200': { description: Item added }
  /cart/items/{id}:
    patch:
      summary: Update cart item quantity
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Item updated }
    delete:
      summary: Remove cart item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Item removed }
  /cart/checkout:
    post:
      summary: Checkout cart
      responses:
        '200': { description: Checkout payload }
        '409': { description: Conflict }